backend/requirements.txt
Django==4.1.2
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
Pillow==9.2.0
psycopg2==2.9.4
